Moved mapToBackend and mapToFronend to common directory
