+++ src/reducers/storageunit/panel/index.js
@@ -1,5 +1,4 @@
-import { parseFloatFromString, formatFloatToString } from '../../../util'
-
+import { mapToFrontend, mapToBackend } from '../mapper'
 const INSERT = 'musit/storageunit-container/INSERT';
 const UPDATE = 'musit/storageunit-container/UPDATE';
 const INSERT_SUCCESS = 'musit/storageunit-container/INSERT_SUCCESS';
@@ -9,48 +8,6 @@ const LOAD_SUCCESS = 'musit/storageunit-container/LOAD_SUCCESS';
 const LOAD_FAIL = 'musit/storageunit-container/LOAD_FAIL';
 
 const initialState = {}
-
-const mapToFrontend = (data) => {
-  return {
-    ...data,
-    area: formatFloatToString(data.area),
-    areaTo: formatFloatToString(data.areaTo),
-    height: formatFloatToString(data.height),
-    heightTo: formatFloatToString(data.heightTo),
-    environmentRequirement: data.environmentRequirement ? { ...data.environmentRequirement,
-      temperature: formatFloatToString(data.environmentRequirement.temperature),
-      temperatureTolerance: formatFloatToString(data.environmentRequirement.temperatureTolerance),
-      hypoxicAir: formatFloatToString(data.environmentRequirement.hypoxicAir),
-      hypoxicAirTolerance: formatFloatToString(data.environmentRequirement.hypoxicAirTolerance),
-      relativeHumidity: formatFloatToString(data.environmentRequirement.relativeHumidity),
-      relativeHumidityTolerance: formatFloatToString(data.environmentRequirement.relativeHumidityTolerance)
-    } : {},
-    environmentAssessment: data.environmentAssessment || {},
-    securityAssessment: data.securityAssessment || {}
-  }
-}
-
-const mapToBackend = (data, parentId) => {
-  return {
-    ...data,
-    groupRead: 'foo', // Must be removed
-    isPartOf: parentId ? parentId * 1 : data.isPartOf,
-    area: parseFloatFromString(data.area),
-    areaTo: parseFloatFromString(data.areaTo),
-    height: parseFloatFromString(data.height),
-    heightTo: parseFloatFromString(data.heightTo),
-    environmentRequirement: data.environmentRequirement ? { ...data.environmentRequirement,
-      temperature: parseFloatFromString(data.environmentRequirement.temperature),
-      temperatureTolerance: parseFloatFromString(data.environmentRequirement.temperatureTolerance),
-      hypoxicAir: parseFloatFromString(data.environmentRequirement.hypoxicAir),
-      hypoxicAirTolerance: parseFloatFromString(data.environmentRequirement.hypoxicAirTolerance),
-      relativeHumidity: parseFloatFromString(data.environmentRequirement.relativeHumidity),
-      relativeHumidityTolerance: parseFloatFromString(data.environmentRequirement.relativeHumidityTolerance)
-    } : {},
-    environmentAssessment: data.environmentAssessment || {},
-    securityAssessment: data.securityAssessment || {}
-  }
-}
 
 const storageUnitContainerReducer = (state = initialState, action = {}) => {
   switch (action.type) {
